trans RDC service from js to ts
